Heapsort
Worst-case performance	 O( n log ⁡ n ) 
The heapsort algorithm can be divided into two parts.  In the first step, a heap is built out of the data. In the second step, a sorted array is created by repeatedly removing the largest element from the heap (the root of the heap), and inserting it into the array.